change scope wizard structure
